static/static/requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==2.0.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.3
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djlint==0.7.6
et-xmlfile==1.1.0
filelock==3.6.0
frozenlist==1.3.0
gunicorn==19.9.0
idna==3.3
import-export==0.2.67.dev6
importlib-metadata==4.11.3
MarkupPy==1.14
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
pipenv==2022.3.24
pipx==1.0.0
platformdirs==2.5.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tomlkit==0.9.2
tqdm==4.63.0
tzdata==2021.5
urllib3==1.26.9
userpath==1.8.0
virtualenv==20.13.4
virtualenv-clone==0.5.7
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.7.0
